feat: ingredients should accept decimal and 0 nutritional values

+++ src/GlucoPilot.Api/Endpoints/Meals/List/ListMealsResponseModels.cs
@@ -16,10 +16,10 @@ namespace GlucoPilot.Api.Endpoints.Meals.List
         public required string Name { get; set; }
         public required DateTimeOffset Created { get; set; }
         public DateTimeOffset? Updated { get; set; }
-        public required int TotalCalories { get; set; }
-        public required int TotalCarbs { get; set; }
-        public required int TotalProtein { get; set; }
-        public required int TotalFat { get; set; }
+        public required decimal TotalCalories { get; set; }
+        public required decimal TotalCarbs { get; set; }
+        public required decimal TotalProtein { get; set; }
+        public required decimal TotalFat { get; set; }
         public List<MealIngredientResponse>? MealIngredients { get; set; } = new();
     }
 
@@ -36,10 +36,10 @@ namespace GlucoPilot.Api.Endpoints.Meals.List
     {
         public required Guid Id { get; set; }
         public required string Name { get; set; }
-        public required int Carbs { get; set; }
-        public required int Protein { get; set; }
-        public required int Fat { get; set; }
-        public required int Calories { get; set; }
+        public required decimal Carbs { get; set; }
+        public required decimal Protein { get; set; }
+        public required decimal Fat { get; set; }
+        public required decimal Calories { get; set; }
         public required UnitOfMeasurement Uom { get; set; }
         public DateTimeOffset? Created { get; set; }
         public DateTimeOffset? Updated { get; set; }
